Initial commit
This patch adds the core panels, dev install script, basic
documentation, and testing.

There is a fairly large amount of extra "stuff" left in, due to the
copy/paste nature of the setup. Specifically, the tests have a lot of
the horizon test code in which is unecessary, but temporary with the aim
of getting up and running quickly.

Change-Id: I5fdce186df09e1168e8d52abfec17e0c20a64031
Partially-Implements: blueprint initial-setup